First working Iteration on IABook reader 4.2 Integreation with Strawberryfield
This pull adds:
Custom JS to use IABinlinookreader as a field formatter, and process either a fixed URL or an inline genertated Manifest. The code is pretty advanced and tries to deal even with some border cases.
It also adds the formatter itself and a Metadata Display Generation Strategy = Means you can use twig template to genertate the Book Manifest.
`@TODO:'
- PDF integration and no manifest (simple image listing) for the IABOOK reader
- Search and integration.
- Documentation