Adds the strawberry_audio_formatter to play audio/podcasts
Works quite similar to any of the other formatters we have. Difference is its using our new IIF route 'format_strawberryfield.iiifav' to generate the Source URL
path: '/do/{node}/iiif/{uuid}/FULL/100/{format}'
This is mostly to allow:
S3:// non Private files to be played in a localhost environment or where access to the file is non public.
Todo: Evaluate https://api.drupal.org/api/drupal/core%21includes%21file.inc/function/file_create_url/8.2.x and add an alter hook. The fact that a file us not publicly available is generating issues with Webform preview of existing assets. This is not an issue with Sites that are published in a real domain on the web.

Also: The new route allows us to allow streaming instead of downloading files. This is another TODO to allow large videos and audio to start playing inmediatelly and will be dealt separately. @giancarlobi there are some things to do here. I need to able able also to play external resources, but that is a change that will happen on all formatters once i publish video, 3D model and data sets later tonight