Use new Dependency format for Drupal
See most recent one
https://www.drupal.org/docs/8/creating-custom-modules/let-drupal-8-know-about-your-module-with-an-infoyml-file

I'm a bit confused abot this one since are not planning on having our modules on the drupal website for now.
Also: format does not have a dependency really on webform, at least on terms of code, but for now its the simplest way of ingesting. Specially now that we moved most of the heavy code to a service in strawberryfield/strawberryfield