Move Twig template processing to Metadata Display Entity
This gives the Entity some processing responsabilities, allowing us to avoid duplicating some code on Controllers and formatters. All this because the moment we want to use a Metadata Displaty Entity to process some JSON into some output, we are already loading the entity. So better let it do its own processing. Also adds base validators for JSONLD and XML. Not in use yet, but comming in the next pull