First working Controller for Metadata Expose Display functionality
This needs some testing. It can handle all our mimetypes and is also caching aware. But i don't know yet is how to deal with the edge case of a SBF field with multiple values (delta piece) of a custom entity where we have multiple fields of SBF type. We don't ship that option and right now its defaulting to the last one.

This also brings rendering to the Metadata Display entity itself and adds the "get used Variables" method we will have fun using.