Restores IIIF General setting form to working condition
This changes also assumptions because we need this to be readable:
if the method is named checkURL then we assume return TRUE means its valid, so the var that gets the result should not be say error, but valid. Also, the "type" was a random string (trying to match the formatter settings form element) that was not documented. Breaking code like this is easy. So we have now 2 class level constants.