Update StrawberryBaseFormatter.php
Give each method proper PHPdocs. Remove unused USE statements, fix wrongly concatenated t() statements and use the correct placeholder replacements.
Change all false and true to FALSE and TRUE and fix the Set Error By name to make sure we are really showing the error inline and not somewhere invisible on the top of the page when setting things. Use the Constants when calling validator and don't use/assume element name as a valid argument, we have now valid class Constants. Lastly, rename an unused (for now) method to correct camelcase (that was left from old code, not introduced here)