Getting local configs wrongly
Local settings can not access the global getter because while you are changing between local and global, the setting that is being used to check what is enabled is not saved yet! Means: i have global settings enabled, i go and check "use locals", i get the new fields to fill up locals, but those values will be the global ones because in the saved settings "use globals" is on.