Remove unused code and options
We are going only for IIIF manifests here, no support for image sequences out of that until JS is actually dealing with that option. Also, Our current IIIF manifest twig template depends on having a / at the end (makes logic simpler) so pass it correctly as part of the $context. WE need to discuss @marlo-longley and @giancarlobi this remove/leave / at the end and come to a concensus. Right now the slash is being removed not from the saved config for locals but on getting the value back. Which means if something accesses the local values directly it will fail.
Also all out twig templates (shipped) need to be synced with this changes.