JSON Type to ADO Mapping
First Automatic View Mode triggering based on the value present in the type keys for Archipelago Digital Objects commit. Provides a Settings form with a draggable table, well defined configuration schema and a hook that alters the view mode, when accesing a NODE's landing-canonical page which was no custom view mode enabled or is tagged to use the "default one".