Show the View Mode label in the Table Drag
Changes the Table Drag Column for View Mode from 'textfield' to 'value' and shows the actual label to be consistent with the select field. This also makes sure people can not afterwards change the View Mode.
I left Type open to be edited since admins could want to use 'types' that don't exist in the system yet.