Check if view mode is existist in current system. If not, allow admin to fix
If during Config load, we find that one setting points to a not found/never existed/was removed View Mode, allow the Admin to fix it by changing the non existing value with a Select Box. Worst case scenario the admin is blind and will be saved using the default view mode.

@TODO. Maybe we can add some checks in the global drupal state check that runs to see if the there are general errors. Also we could remove a setting when a View Mode is deleted?