Pass the rest of the manifests to the list so people can add new Windows based on that
This requires a TODO: The correct structure is
"THEURL":{'provider':'some institution'}.
We are not yet passing that data from the SBF to the JS. That needs to be worked out. We are assuming the data coming from the JSON itself is just a list of URLs or our won. Attribution is IMPORTANT!.