Weirdest issue i have seen. PHP 7.3 (at least our PHP-FPM on alpine) BUG
Foreach does not change its =>$value if the value previously assigned was a string and then a 0. This is just for comparison needs (means if you do a $value == or a switch case: > But if you print the value, or as i do here, reassign it to another value, then comparison works again.

I need to test this in pure PHP and see if this is just our case or anyone has seen this.

SO WEIRDD. I left $summary[] as an ugly concatenated ifs. Seems to work. But a switch would have made the same really. It was the foreach/never the switch.