ISSUE-49: Invite Mirador IIIF Viewer 3 to have dinner with the family (#50)
* Mirador JS stuff

CDN, simple model (will extend in another pull to make this like i wrote it in pieces!)

* New formatter and utility class for the MetadataExposeConfig Entity

Formatter with complex Config form. Will after this add the logic to push the other 2 sources into the JS.

* Unify iiif_server key passed to Twig template

And remove a string casting. Not an issue really in PHP 7.3, but something was complaining about default value and string strong typing.. Should revisit. Also adding 'access-control-allow-origin' = * because we love to SHARE!

* Make Twig Formatter extended from the base class

* strawberry_mirador_formatter schema

Hope i got it right. Since we have some new settings, i need to update the main rendered still to react to those. But this should match the settings form

* Handle 3 different sources for IIIF manifests and cleanup settings

Better settings. Debugged some interaction and ajax is working perfectly
-3 options for sources with correct logic to handle them.

Next step, modify JS to deal with 3 sources and add a secondary screen to Mirador if more than a single panorama is present.

* And i got it. Multi Manifest, from 3 sources working

And if you have more than one, you get two windows

* Recalculate main source of IIIF based on what is checked

This makes more sense. If current main source is un-checked by the user, the default needs to reset itself. Also change the label

* Pass the rest of the manifests to the list so people can add new Windows based on that

This requires a TODO: The correct structure is
"THEURL":{'provider':'some institution'}.
We are not yet passing that data from the SBF to the JS. That needs to be worked out. We are assuming the data coming from the JSON itself is just a list of URLs or our won. Attribution is IMPORTANT!.

* Remove empty statements

* Make sure we have a second manifest before showing a second window

* Weirdest issue i have seen. PHP 7.3 (at least our PHP-FPM on alpine) BUG

Foreach does not change its =>$value if the value previously assigned was a string and then a 0. This is just for comparison needs (means if you do a $value == or a switch case: > But if you print the value, or as i do here, reassign it to another value, then comparison works again.

I need to test this in pure PHP and see if this is just our case or anyone has seen this.

SO WEIRDD. I left $summary[] as an ugly concatenated ifs. Seems to work. But a switch would have made the same really. It was the foreach/never the switch.