ISSUE-59: Allow 100% with for all formatters and some ++candy (#60)
* First pass on reverting bugs introduced by Pull 19

For some strange reason, this schema keys wen't this weird route. This is the first step on fixing these. Introduced in #19

* More fixes. This is moving fast and good

* identation!

* We have mixed values. Some are string, some are ints

Figure out what the form is really saving right?

* First pass on normalizing with to use '0' as shortcut for 100%

This touches every formatter. Should not require any changes on current settings.

* Allow 0 heigth on PDF to force proportional height

* Mixed width and height on openseadragon

reacting to my lack of sleep

* Well, canvases need fixed sizes

Will add an onresize callback after this commit. Fix to 720 px as starting point if you have 0 as width. JS will then do a full rescale to scale to pixels based on the parent container

* First shot on canvas resize...

* globalize the local

* We know canvas is there

* Jquery-ize

* log/debug

* Another approach

* move viewer.FitInWindow(); to after

* set attr and css

* And now it works

Next step, the base64 encoded snapshot of the 3D rendering to allow a Thumbnail.

* Try a loading notification

* use D8 std classes for ajax progress

* Try own css

* loader centered via absolute css

Kinda 90's but better than a white screen while the world changes

* ups.jquery.ups

* wrong selector, but better via the class anyway

* Let the preloader start a little bit earlier. Let it go if something fails

* For fun.. will delete after

* Cross domain messed the fun

Lets try something local

* Add model/obj to our list and remove test

Doing this directly on formatter but should on beta3 move to SBF?

* Simpler summary for width and height...

Since we have defaults now and also 0 evaluates to false.
Also make width and heigth and JSON KEY required elements.

* Worst copy pasta ever

Ok, now it should be Ok again

* Same mistake i made 5 hours ago

Should i sleep or more coffee?

* And again...

* Pass no widht and height attribute.. this is not 1996

* Allow 0 height for Video as stub for auto/proportional

Because squeezed video is so 1998

* Forgot we have two processors for paged. Done!

* Thanks to Giancarlo super eyes!

Also, i just found i was issuing a getSettings (see the s!) instead of singular for the thumbnails. That is not right and was adding a lot of info to that data- property in the HTML. FIXED!