IABookreader IIIF fix: Uses service id only if present, passes arguments over
This commit, for IIIF Display API, item manifests 2.1,  uses service @id only if present and passes any arguments the image could have to the Image URI builder (if not uses the given Resource/Image @ID as instructed).

I'm still doubtfull this is the correct way, i feel we should always, like always use the Image ID. This current behaviour was inherited from the demo code that IABookreader had for IIIF, and it feels wrong.  Mostly because, if the IIIF manifest is already pointing to a given Image URL, then making an aditional call to a (maybe not even existing) info.json is overriding the users/creator wishes. Again, the original code was not done by us.

In any case, i need to also remove any calls to with and height if there is no service right? Maybe we need to test with some IIIF manifest (public?) where service is not present. @giancarlobi