Adding some options, now without bar/with bar
This is quite simple. If there is a starting URL, we provide it and the bar goes away. If not, we check if bar is enabled and show it. If not, we just show 'pages' selector so the user can decide. Its late and i'm tired, there can be better options for realz