Issue 74b: IABookreader, please cooperate with old friend PDF? (#83)
* Use Canvas Aspect ration to deduce missing resource dimension

I think a lot of folks have missread the IIIF API Presentation V2.1 specs. Canvas width and height are there for aspect ration. Means we can deduce one of the missing dimensions of a resource based on that. I will try to do that. This is all for LEVEL0 service (no info.json)

* Also notify the right thing

* Some further JS goodness

I wish i knew how to document this!

* Honestly? A typo?

C'mon

* Aspect ratio needs to divide if going height based on width

* Remove console.log

* And the last one