Properly delete Annotations
And to make sure all works, also do an array_values to maintain the proper JSON indexed array structure and i also clean the "main image" key once empty. If not we end with objects with keys like "1". @giancarlobi ready for you to test. I also now simply redirect people out of the page and back to the edit page if they discard. Only safe way i have found to really reload the page and restore the old annotations!