Testing huston.
I do not know if i really? need to be doing Garbage collection here since our errors were mostly related to a Bug in PHP 7.1,7.2 and 7.3 (funny all platforms except el capitan 7.3.9, probably because of a different CURL library).

This is still exploratory. Streaming quicks in only if the size is larger tan 200M and there is not RANGE header request. Or if its forced via an GET argument. Quite some options.