Adds super simple lazyload
I had to "not" use once() because views can load data via ajax and once that happens the observer is not called again. We may want to review this a few times but it works.
In a twig template we can use this

<img class="field-iiif image-iiif img-fluid rounded iiif-lazy" id="thumb_iiif-field_descriptive_metadata-{{ node.uuid.value }}-0-image1" data-src="{{ source }}" width="240" alt="Thumbnail for {{ data.label }}" typeof="foaf:Image">
<noscript>
<img src="{{ source }}" class="field-iiif image-iiif img-fluid rounded" id="thumb_iiif-field_descriptive_metadata-{{ node.uuid.value }}-0-image1" width="240" alt="Thumbnail for {{ data.label }}" typeof="foaf:Image">
</noscript>

If JS is disabled at least the normal thumbnails will load. Issue is with "alt" tag that adds an undesired "broken" image while loading.I tested simulating a slow 3G connection and it did what i needed to do.