See ISSUE-138: quick fix while i review the cache tag complexity
After a lot of testing invalidating the complete cache tag of a Node for data that is   not part of the Node itself (means the NODE and all its cached data is valid) was very performant breaking. For now, while i review Pedro's large cache invalidating Pull i will move with Ajax/JS for the CSS hiding of the search box.  Depends on the lates 1.0.0-RC2 Strawberryfield and its new route for counting docs