Fixes Mobile IIIF image for Panellum
@giancarlobi this fixes your issue
We now pass both versions of the image and on JS decide which one to use.
This will make sure that during the whole tour a 4K version max is used for mobile (conservative but cross compatible with most phones).

You need to clear your caches and also browser caches since JS changed a bit!