Returns actually results
This also required a change in the Twig Template preview. Asked community. Do we want to preview as anonymous? Or as logged in user?

For some reason when  $this->renderer->renderPlain($build); runs on a validation hook the current user gets lost! and all from there on renders as anonymous. We have been doing this since for ever but i know feel it is weird. This gets around it but i see how rendering as anonymous (preview) might have its benefits?