Remove base64 encoding, just makes all larger
Add grouping, not sure if its working well, but it looks like, will probably debug a bit. Each grouped element (e.g Parent Collection) generates a new Layer that is also bound to the Marker Cluster. Also removed redundant options for the Style Plugin. This still needs a schema yml for the config