Fix Ajax and add multi view UI. Never set IDs manually for Submit buttons that trigger AJAX
I have repeated this to myself for the last 3 years and i always forget. I even have a sticky note
Given that Sets and Items have to be handled by different views we need to enable Multiple Views that can be modal based on enumeration properties.
I'm pretty sure i coded this for my birthday already but can't find it. Too many branches so here we go again