Basic pass custom Mirador 3.3.0 with Image tools
I could not make the awesome text-overlay the Johannes wrote play with image tools that community asked for. When both are enabled image tools are not drawn. I have no idea what is happening and that will imply extra work. Some funkyness with opening new manifests when pressing on the blue signs are also happening, might be some material-layour issue, a css that is missing, a selector or even the URL parser. JS -react- are a mistery for me