basically remake ajax facets
So wrongly coded. This JS should stick better. Also big Ajax Views error, if the normal exposed form is not there and you return still the replace selector all dies, Drupal.views ajax fails to initialize and a race condition happens (after a single Facet click, even without our code). I fixed this but check if the selector is in place before adding our ajax handlers and also checking on that on the ajax controller we use as replacement. Hopefully this works out.