Adds the option to "not join" in case of a negation
WIP also a comma separated list of processor types. This could have been a facet with options, etc.. But i feel this approach is more simplistic. Will now add this to the query itself AND work on the "exposing" (basically enabling/disabling OCR search via a checkbox if the user decides to expose this)