Normalize properties/inline doc here (READ THIS)
For an ADO change to be triggered it needs:
"sbf:ado:change": nid
for an view change the JSON-LD or the GEOJSON needs
'sbf:ado:view:change': nid or any exposed/relationship value one wants to pass/view receiving accepts.
For a canvas change it needs (Mirador does nothing, it will trigger the canvas change on change) but maps needs

"sbf:ado:canvas:change": {
   "canvasid": the id of the canvas,
  "manifestid": the id of the manifest
}

This is a bit tooo precise for my taste? mmmm
Wonder if Mirador should also react to ADO change? And i do the "find the canvas that contains the actual ID i'm searching for? mmmm
Probably