Still a few hours here? but this is now mostly working
What is not YET working?
The View is caching old Submit values. Means when i delete one and re-add one, the old "search" term reappears ... as a default. Who knows why? I will never know why. But i basically need to intervene either here or in the Ajax Controller to avoid non-accepted/not-parsed values to be take by the Filter as valid/kept around...