Be more carelfully when dealing with sbf_ajax_interactions after disabling ajax
Now we will show/but disabled on the absence of "use Ajax" but also clear if after setting this up, Ajax is disabled. Also made sure the wording is accurate. Right now we process Contextual filter values/exposed on path values. Not form driven exposed arguments (imagine the mess if we did that)