If the path is wrong target_url == NULL
Bc views are a mess. I messed up a path, which then did not longer was valid, and because of that Ajax itself died. That is actually OK BC drupal should not even allow a Path to be visible if it is invalid! (according to its own validation methods) but bc it allows it to be there (caching, who knows) at least i can avoid generating any GET/URL/Browser history if that is the case