Adds a pseudo Node Upcaster for Views
This tries to get around the exception of a View using %node as a parameter but not upcasting it to NodeInterface. It, of course, does this without too much overhead, which means it does not processes the View  Entity Itself and simply assumes a view_id + node argument are reasons enough to enrich the Route. It is important to node that really only place where this can happen is when a View's path is node/%node given that (fixed in code) that is where our Route (Active Display settings) will call this class. Basically when Node level Tabs (actions) and View level Actions depend on each other