Need to commit bc this stuff is hard ... so hard
Basically i got the property solving for ::calculateEntityRelationsForField wrong.
The right approach (which is not here yet is)..
Iterate from root to leaves all property paths. IF i find an entity relationship, reset the accumulated property paths from there (basically tread the foreign relationship as one that needs to be resolved afterwards locally (new root) agains the filter value. Keep doing that. Everytime that happens, we reset again. Don't accumulate $data... only keep one around and it needs to be filled up AT the end. Once we run out of paths.
- IN effect this will
  - Preserve directly resolvable paths (e.g field_descriptive_metadata:label or title directly -> good for comparing ADOS against each other
 - Provide a clean Property path in case of a complex one. e.g 0 = field_descriptive_metadata:sbf_entity_reference_ismemberof:field_descriptive_metadata:digital_object_type would be just field_descriptive_metadata:digital_object_type
(this is me explaining this to myself bc it is hard... gosh... why do i code this stuff?)