Much better. But we need to improve also the entity reference plugin
In webform_strawberryfield. Don't forget @diegopino @alliomeria (remind me?). Maybe the solution there to avoid "select referencing" webform elements (the Collection autocomplete that shows itself in the list?) would be to pass as an extra option/argument to the Autocomplete itself via webform ? the current NODE id so we can exclude it?