Probably a better way of detecting multiple nested for/endfor
With some having if inside and some not.
By keeping track of the "level" we are in, and decreasing deepness on endfors found we can do this in a more reliable way without having to keep track if found or not as a single dimensional "reality" @alliomeria this fixes the AMI template that you have (which is SUPER complex). More testing Monday maybe?