This adds also Exposed Metadata Endpoints (so indirect usage) on View Modes
Basically. We calculate first if an exposed view mode uses this template, we remember which ones, then on the View Mode settings we check if also that exposed endpoint is used. And now we remember also the View Mode. Why? For the next pull, where we get now Views/Blocks that Might be also using "rendered" entities and thus, also indirectly via these making use of the template.