Adds new config options for the classic mode
We don't act on these yet. On classic mode the default behavior (invisible to the user) will be to render all allowed filters BUT hide them initially. That way the plus and minus signs really only do "show/hide" and unset values instead of depending on a complex logic on the Actual exposed form handler for the filter. This also allows the Filter still have full control of how many/naming, etc. All this via JS