First pass on ADO to ADO filter
I need to still figure out if "flexibility" here is more important than speed. Why?
I can save (as in the code) the Solr Field name exactly the way i need if for the JOIN (or multiple) by allowing the Settings to use that instead of the Search API one, but... if the user decides to re-index an already setup field by changing the type the filter will break. But also, having to calculate the proper Solr name on every query is expensive. What will it be, what @alliomeria ?