mmm.. loooks like we need to deep merge the settings on the facet ajax
controller. What i don't like about this override is that we have to "build" the Facets again, even if they were built before by the parent one. That is an overkill in processing. Probably better? (worse code wise though) is to actually copy the code from the parent method and do our thing here without depending on the parent one ... That is N number of facets X Render less processing @alliomeria but with the difficulty that if Facets decides to change their approach completely i have to re-do all the code in a next release... speed v/s maintainability?