Gosh Drupal. Drupal settings/Ajax page state and Libraries with Facets issue
Ok, this finally solves the warnings/missing data and also has less processing.
because of how Drupal "deep" merges Settings, passing chart data as an array ended in new data/existing data being replaced but all data still visible.
Even if that "might" be desired it is not dependable. So now we JSON encode and then JSON decode via JS to pass those Drupal settings around. This also simplifies the SBF Facet Block Ajax Controller since we don't need to pre-render (like everyone else is doing, we can just pass the Render Array and let the Attachments Aware Trait fetch the dependent libraries and settings. No more dialog.settings.close warning.

Also. On existing ajax_page_state, we pass it as an request variable and remove it from the query. If not a widget's RESET or link will show that whole mess and break the theme