This is complex. Refactor the embargo service and interface
We need to statically cache as UUID and NID the results (in a single call) to avoid over processing. The reason we need the NID (and new method for that) is that at the View level, where we have our new Argument Validator, the argument is a NID .... Also.
When a View checks/validates the argument and e.g We find it is "FALSE", meaning the user can't bypass an embargo, basically no more processing happens and the View might return FALSE (not an render array). IN that case, no more "hooks" run and i can't longer alter the Cache tags/ max times and context. So the safest option for that is really, simply NOT cache at all.

It is one or the other: safety v/s speed. First one wins in the world of Bots/access restrictions.