This change IS really not needed. I am just adding extra "cache" variations just in case
In the past i had only add 'ip' variation to cache IF we had an embargo by IP, but then, i already KILL the cache completely when that happens. means there is no variation. I just felt it was better (memory/old man) to have a single way of doing it everywhere (had mixed code) and the safest way is always "normalize against the more exagerated/security wise way) @alliomeria