Note about removing Time Zone offset from Facet results.
What? We add the offset (or remove) during query to match "input" v/s "actual indexed data". All good there, but once we get the FACETs back, those ranged/buckets have the offsets. So if we use those as input again and YET again calculate offsets, we are basically moving hours back and forth and WE CAN (on the $max) end escaping/reducing our actual range.
Solution for the weekend is re-verse the operation we do on query, so our facets are YET again UTC