Matches escape strategy of D11, dealing differently with TwigMarkup(bypass) and MarkupInterface(cast to string) + sbf_render
sbf_render is a filter, to be used as replacement to |render (for now while https://www.drupal.org/project/drupal/issues/3554880#comment-16326144 goes through/discussions/could take a long time or lead to nothing) that solves the issue when working with Views/Field outputs and rewriting that |render effectively will not act at all on a replacement value from another field passed around as a safe MarkupInterface implementing Object, because and |render will be basically like nothing happened. That is a huge issue/large change from D10, if you are e.g setting a variable to re-used it as argument for another function.