Fixes Twig 3.9 deprecation *obscured somewhere
todo: env->getFilter is an internal function now (marked as such). D11.x is still doing this and we are overriding drupal so it does not crash our User pages if someone decides (by mistake) to output {{ myarray }}. But not sure i can't go my own lengths of creating an Instance level TwigFilter Object in the constructor here? Of get all Filters, and then fetch them by name? Lots of deprecations in Twig